Add test for an edge case where a oneof is named "result"
When a oneof is named "result" (or possibly any other Elm prelude name),
the code generator tries to add underscores at the end of the name to
avoid conflicts. However, sometimes in longer path names in
Internals_.elm the underscore is added after a period, not before as it
should (e.g. Foo.Result._Code vs Foo.Result_.Code).